requirements.txt
absl-py==1.0.0
alembic==1.7.7
asgiref==3.5.0
astunparse==1.6.3
backports.zoneinfo==0.2.1
bleach==5.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.28
Django==4.0.4
dnspython==2.2.1
docutils==0.18.1
easyocr==1.4.2
email-validator==1.2.0
ffmpeg==1.4
Flask==2.1.2
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.18.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.2
libclang==14.0.1
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplot==0.1.9
matplotlib==3.5.1
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
pkginfo==1.8.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyloco==0.0.139
pyparsing==3.0.8
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
PyWavelets==1.3.0
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==35.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.3.0
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
SimpleWebSocketServer==0.1.1
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tesseract==0.1.3
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.4.28
torch==1.11.0
torchvision==0.12.0
twine==4.0.0
typing==3.7.4.3
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.14.0
WTForms==3.0.1
zipp==3.8.0
